INS-3373 move log reader to testutils
